web/requirements.txt
Flask==1.1.1
flask-restplus==0.13.0
psycopg2==2.8.3
python-dotenv==0.10.3
SQLAlchemy==1.3.8
